refactor scoot.ir package name to scoot.jimple
