+++ src/main/scala/edu/colorado/plv/cuanto/scoot/arithmetic/Interpreter.scala
@@ -1,7 +1,7 @@
 package edu.colorado.plv.cuanto.scoot
 package arithmetic
 
-import edu.colorado.plv.cuanto.scoot.ir._
+import edu.colorado.plv.cuanto.scoot.jimple._
 import scala.collection.immutable.HashMap
 
 /** Implement an interpreter for sequences of Jimple assignment
