+++ src/main/scala/edu/colorado/plv/cuanto/scoot/jimple/AddExpr.scala
@@ -1,4 +1,5 @@
-package edu.colorado.plv.cuanto.scoot.ir
+package edu.colorado.plv.cuanto.scoot.jimple
+import soot.jimple
 
 /**
   * Created by Jared on 6/21/2017.
@@ -7,6 +8,6 @@ class AddExpr(dt: soot.jimple.AddExpr) extends BinopExpr(dt: soot.jimple.BinopEx
 //empty interface
 
 object AddExpr {
-  def unapply(e: AddExpr): Option[(Value, Value)] =
+  def unapply(e: AddExpr): Option[(soot.Value, soot.Value)] =
     Some((e.op1, e.op2))
 }
