+++ src/main/scala/edu/colorado/plv/cuanto/scoot/jimple/BinopExpr.scala
@@ -1,4 +1,4 @@
-package edu.colorado.plv.cuanto.scoot.ir
+package edu.colorado.plv.cuanto.scoot.jimple
 
 /**
   * Created by Jared on 6/13/2017.
