+++ src/test/scala/edu/colorado/plv/cuanto/scoot/arithmetic/ArithmeticInterpreterSpec.scala
@@ -3,7 +3,7 @@ package arithmetic
 
 import org.scalatest.{FlatSpec, Matchers}
 import org.scalatest.prop.PropertyChecks
-import edu.colorado.plv.cuanto.scoot.ir._
+import edu.colorado.plv.cuanto.scoot.jimple._
 
 class ArithmeticInterpreterSpec extends FlatSpec with Matchers with PropertyChecks {
   import Builder._
